+++ components/switch/mdc-switch.vue
@@ -7,22 +7,24 @@
     class="mdc-switch-wrapper" >
 
     <div
-      :class="{'mdc-switch--disabled': disabled }"
+      :class="classes"
+      :styles="styles"
       class="mdc-switch">
-      <input
-        ref="control"
-        :name="name"
-        :id="vma_uid_"
-        :checked="checked"
-        :disabled="disabled"
-        type="checkbox"
-        class="mdc-switch__native-control"
-        @change="onChanged" >
+      <div class="mdc-switch__track" />
+      <div class="mdc-switch__thumb-underlay">
+        <div class="mdc-switch__thumb">
+          <input
+            ref="control"
+            :name="name"
+            :id="vma_uid_"
+            :value="value"
+            type="checkbox"
+            role="switch"
+            class="mdc-switch__native-control"
+            @change="onChanged" >
 
-      <div class="mdc-switch__background">
-        <div class="mdc-switch__knob"/>
+        </div>
       </div>
-
     </div>
 
     <label
@@ -37,6 +39,8 @@
 
 <script>
 import { DispatchFocusMixin, VMAUniqueIdMixin } from '../base'
+import { RippleBase } from '../ripple'
+import MDCSwitchFoundation from '@material/switch/foundation'
 
 export default {
   name: 'mdc-switch',
@@ -47,24 +51,57 @@ export default {
   },
   props: {
     checked: Boolean,
+    disabled: Boolean,
+    value: String,
     label: String,
     alignEnd: Boolean,
-    disabled: Boolean,
-    value: {
-      type: String,
-      default() {
-        return 'on'
-      }
-    },
     name: String
+  },
+  data() {
+    return {
+      classes: {},
+      styles: {}
+    }
   },
   computed: {
     hasLabel() {
       return this.label || this.$slots.default
     }
   },
+  watch: {
+    checked(value) {
+      this.foundation && this.foundation.setChecked(value)
+    },
+    disabled(value) {
+      this.foundation && this.foundation.setDisabled(value)
+    }
+  },
+
+  mounted() {
+    this.foundation = new MDCSwitchFoundation({
+      addClass: className => this.$set(this.classes, className, true),
+      removeClass: className => this.$delete(this.classes, className),
+      setNativeControlChecked: checked =>
+        (this.$refs.control.checked = checked),
+      isNativeControlChecked: () => this.$refs.control.checked,
+      setNativeControlDisabled: disabled =>
+        (this.$refs.control.disabled = disabled),
+      isNativeControlDisabled: () => this.nativeControl_.disabled
+    })
+    this.foundation.init()
+    this.foundation.setChecked(this.checked)
+    this.foundation.setDisabled(this.disabled)
+
+    this.ripple = new RippleBase(this)
+    this.ripple.init()
+  },
+  beforeDestroy() {
+    this.foundation && this.foundation.destroy()
+    this.ripple && this.ripple.destroy()
+  },
   methods: {
     onChanged(event) {
+      this.foundation && this.foundation.handleChange()
       this.$emit('change', event.target.checked)
     }
   }